Split Maine SSP living arrangement into input + payment_category
Mirror Alabama SSP pattern: bare living_arrangement enum input plus a
derived payment_category that applies the federal MEDICAL_TREATMENT_FACILITY
override. Replaces the prior _input/derived rename with clearer naming.
