+++ policyengine_us/parameters/gov/states/me/dhhs/ssp/amount/individual.yaml
@@ -4,7 +4,7 @@ metadata:
   period: month
   label: Maine SSP individual monthly amount
   breakdown:
-    - me_ssp_living_arrangement
+    - me_ssp_payment_category
   reference:
     - title: 10-144 C.M.R. Ch. 332, Part 11, Section 5 (MaineCare Eligibility Manual; PDF page 127)
       href: https://www.maine.gov/sos/sites/maine.gov.sos/files/inline-files/144c332-2025-101%20%28AMD%29_0.docx
