+++ policyengine_us/tests/policy/baseline/gov/states/me/dhhs/ssp/integration.yaml
@@ -38,7 +38,7 @@
         age: 50
         is_ssi_disabled: true
         employment_income: 0
-        me_ssp_living_arrangement_input: ADULT_FOSTER_HOME
+        me_ssp_living_arrangement: ADULT_FOSTER_HOME
     tax_units:
       tax_unit:
         members: [person1]
@@ -70,13 +70,13 @@
         is_ssi_aged: true
         employment_income: 0
         ssi_claim_is_joint: true
-        me_ssp_living_arrangement_input: FLAT_RATE_BOARDING_HOME
+        me_ssp_living_arrangement: FLAT_RATE_BOARDING_HOME
       person2:
         age: 70
         is_ssi_aged: true
         employment_income: 0
         ssi_claim_is_joint: true
-        me_ssp_living_arrangement_input: FLAT_RATE_BOARDING_HOME
+        me_ssp_living_arrangement: FLAT_RATE_BOARDING_HOME
     tax_units:
       tax_unit:
         members: [person1, person2]
@@ -107,7 +107,7 @@
         age: 75
         is_ssi_aged: true
         social_security_retirement: 6_000
-        me_ssp_living_arrangement_input: RESIDENTIAL_CARE_FACILITY
+        me_ssp_living_arrangement: RESIDENTIAL_CARE_FACILITY
     tax_units:
       tax_unit:
         members: [person1]
@@ -167,7 +167,7 @@
         age: 70
         is_ssi_aged: true
         employment_income: 0
-        me_ssp_living_arrangement_input: ADULT_FAMILY_CARE_HOME
+        me_ssp_living_arrangement: ADULT_FAMILY_CARE_HOME
     tax_units:
       tax_unit:
         members: [person1]
@@ -196,7 +196,7 @@
         age: 60
         is_ssi_disabled: true
         employment_income: 0
-        me_ssp_living_arrangement_input: COST_REIMBURSED_BOARDING_HOME
+        me_ssp_living_arrangement: COST_REIMBURSED_BOARDING_HOME
     tax_units:
       tax_unit:
         members: [person1]
