+++ policyengine_us/tests/policy/baseline/gov/states/me/dhhs/ssp/me_ssp_eligible.yaml
@@ -73,7 +73,7 @@
     people:
       person1:
         ssi: 600
-        me_ssp_living_arrangement_input: ADULT_FOSTER_HOME
+        me_ssp_living_arrangement: ADULT_FOSTER_HOME
     households:
       household:
         members: [person1]
