+++ policyengine_us/variables/gov/states/me/dhhs/ssp/me_ssp.py
@@ -19,51 +19,13 @@ class me_ssp(Variable):
         joint_claim = person("ssi_claim_is_joint", period.this_year)
         eligible = person("me_ssp_eligible", period)
         both_eligible = person.marital_unit.sum(eligible) == 2
-        arrangement = person("me_ssp_living_arrangement", period)
-        values = arrangement.possible_values
-        marital_unit_size = person.marital_unit.nb_persons()
-        # Both spouses must share the same living arrangement to be
-        # treated as a joint couple claim. New enum values added to
-        # MESSPLivingArrangement must be appended to this check.
-        shared_arrangement = (
-            (
-                person.marital_unit.sum(
-                    arrangement == values.LIVING_ALONE_OR_WITH_OTHERS
-                )
-                == marital_unit_size
-            )
-            | (
-                person.marital_unit.sum(arrangement == values.HOUSEHOLD_OF_ANOTHER)
-                == marital_unit_size
-            )
-            | (
-                person.marital_unit.sum(arrangement == values.ADULT_FOSTER_HOME)
-                == marital_unit_size
-            )
-            | (
-                person.marital_unit.sum(arrangement == values.FLAT_RATE_BOARDING_HOME)
-                == marital_unit_size
-            )
-            | (
-                person.marital_unit.sum(arrangement == values.ADULT_FAMILY_CARE_HOME)
-                == marital_unit_size
-            )
-            | (
-                person.marital_unit.sum(
-                    arrangement == values.COST_REIMBURSED_BOARDING_HOME
-                )
-                == marital_unit_size
-            )
-            | (
-                person.marital_unit.sum(arrangement == values.MEDICAID_FACILITY)
-                == marital_unit_size
-            )
-            | (
-                person.marital_unit.sum(arrangement == values.RESIDENTIAL_CARE_FACILITY)
-                == marital_unit_size
-            )
+        category = person("me_ssp_payment_category", period)
+        # Both spouses must share the same payment category to be treated
+        # as a joint couple claim.
+        shared_category = person.marital_unit.max(category) == person.marital_unit.min(
+            category
         )
-        couple_applies = joint_claim & both_eligible & shared_arrangement
+        couple_applies = joint_claim & both_eligible & shared_category
         return where(
             couple_applies,
             person("me_ssp_couple", period),
